+++ R/load_last.R
@@ -21,6 +21,12 @@ load_last <-function(file) {
 		stop("The requested backup is missing in the working directory.")
 	Name <- sub(".rda", "", inFolder, fixed=TRUE)
 	Name <- strsplit(Name, "_", fixed=TRUE)
+	step_back <- function(x) {
+		if(length(x) > 2) x <- c(paste(x[1:(length(x) - 1)], collapse="_"),
+					x[length(x)])
+		return(x)
+	}
+	Name <- lapply(Name, step_back)
 	Name[sapply(Name, function(x) length(x) == 2)] <- lapply(Name[sapply(Name,
 							function(x) length(x) == 2)], function(x) c(x, "0"))
 	Name <- as.data.frame(do.call(rbind, Name), stringsAsFactors=FALSE)